[FIX] credal_identification partner form view: remove partner_id field from ID numbers form and tree views included in the res.partner form view.
Up to now the partner_id field was visible and filled with active_id.
The problem is that when creating a new partner from the interface
and directly creating new ID numbers before saving
(and thus creating the partner in DB), the active ID
doesn't exist yet and the partner_id couldn't be filled.